+++ core/src/uy/com/agm/gamethree/sprites/items/powerups/PowerOne.java
@@ -132,7 +132,9 @@ public class PowerOne extends Item {
         }
 
         if (stateFadingTime > FADING_SECONDS) {
-            world.destroyBody(b2body);
+            if(!world.isLocked()) {
+                world.destroyBody(b2body);
+            }
             currentState = State.FINISHED;
         }
     }
@@ -140,7 +142,9 @@ public class PowerOne extends Item {
     @Override
     protected void stateTaken(float dt) {
         // Destroy its b2body
-        world.destroyBody(b2body);
+        if(!world.isLocked()) {
+            world.destroyBody(b2body);
+        }
         applyPowerOne();
         currentState = State.FINISHED;
     }
