+++ core/src/uy/com/agm/gamethree/sprites/weapons/hero/HeroBullet.java
@@ -97,7 +97,9 @@ public class HeroBullet extends Weapon {
 
     @Override
     protected void stateOnTarget(float dt) {
-        world.destroyBody(b2body);
+        if(!world.isLocked()) {
+            world.destroyBody(b2body);
+        }
         setBounds(getX(), getY(), Assets.getInstance().getHeroBullet().MUZZLE_FLASH_WIDTH_METERS, Assets.getInstance().getHeroBullet().MUZZLE_FLASH_HEIGHT_METERS);
         setRegion(muzzleFlashImpactFX);
         currentState = State.IMPACT;
